Update Dart compiler and machine outputs

+++ tests/machine/x/dart/fun_three_args.dart
@@ -1,5 +1,5 @@
 int sum3(int a, int b, int c) {
-  return a + b + c;
+  return ((a + b as int) + c as int);
 }
 
 void main() {
