+++ tests/machine/x/dart/let_and_print.dart
@@ -3,5 +3,5 @@ var a = 10;
 int b = 20;
 
 void main() {
-  print(a + b);
+  print((a + b as int));
 }
