+++ tests/machine/x/dart/pure_global_fold.dart
@@ -1,7 +1,7 @@
 var k = 2;
 
 int inc(int x) {
-  return x + k;
+  return (x + k as int);
 }
 
 void main() {
